educations/goit_school/home_works/python_web/home_work_14/contact/pyproject.toml
[tool.poetry]
name = "contact"
version = "0.1.0"
description = ""
authors = ["VadimTrubay <vadnetvadnet@ukr.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.97.0"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
sqlalchemy = "^2.0.16"
psycopg2 = "^2.9.6"
alembic = "^1.11.1"
pydantic = {extras = ["email"], version = "^1.10.9"}
passlib = "^1.7.4"
python-multipart = "^0.0.6"
libgravatar = "^1.0.4"
faker = "^18.13.0"
fastapi-mail = "^1.3.1"
python-dotenv = "^1.0.0"
redis = "^4.6.0"
fastapi-limiter = "^0.1.5"
git-filter-repo = "^2.38.0"
cloudinary = "^1.33.0"
pytest = "^7.4.0"
pytest-mock = "^3.11.1"
pytest-cov = "^4.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}


[tool.poetry.group.dev.dependencies]
sphinx = "^7.0.1"


[tool.poetry.group.test.dependencies]
httpx = "^0.24.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
